fix(lint): extract codex-resume-hint to separate file to fix max-lines violation
- Extracted `renderEntrypointCodexResumeHint` and `escapeForDoubleQuotes` from
  `codex.ts` into `codex-resume-hint.ts` to bring `codex.ts` under 300-line limit
- `codex.ts` re-exports `renderEntrypointCodexResumeHint` to preserve public API
- WHY: ESLint max-lines rule requires files <= 300 lines; codex.ts exceeded limit
  after adding longcontx profile config in feat(config) commit
